+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Multiplicative method for Holt-Winters time series prediction.
+### Changed
+- Holt-Winters no longer additive by default, must specify a method, either `additive` or `multiplicative` in the Holt-Winters configuration.
 
 ## [v0.2.0] - 2019-12-19
 ### Added